Training Management System C Sharp
- This project was developed to bring online various activities
required by a training institute.
- It provides electronic security of documents, is user friendly and
all the database is maintained in an organized way.
- There is personal account for each department which includes the
basic information of the student, teacher, administration, career
counselor and director of the institute.
- The teachers and students can upload and download assignments on the
routine basis.
- It manages all the entries regarding the students who have registered
for the training in which they can choose the language of their own
interest and they are allotted a batch and the teacher.
- The teachers are allotted the batches according to their
qualification.
- After the completion of the training a project is submitted by the
students and is certified accordingly.